requirements.txt
#Python Generic
python-dotenv==1.0.0
chardet==5.2.0

#Open AI
openai==1.11.0

#Langchain
langchain==0.1.8
langchain-community==0.0.25
langchain-experimental==0.0.52
langchain-openai==0.0.5
langgraph==0.0.24
langchainhub==0.1.14

#Streamlit
streamlit==1.31.1
# azure
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-search-documents==11.4.0